2019/general/chainerrl/requirements.txt
chainerrl>=0.7.0
minerl==0.2.3
pydot
matplotlib
